requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic[email]==2.11.7
sqlalchemy==2.0.36
fastapi_sqlalchemy==0.2.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-decouple==3.8
emails==0.6
httpx==0.27.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
alembic==1.13.3

# Development dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.12.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0